Created Physician and Patients User Data

+++ src/com/company/Entity/Patient.java
@@ -3,7 +3,7 @@ package com.company.Entity;
 public class Patient extends User{
     private long patientId;
 
-    public Patient(String name, String address, int age,int mobile) {
+    public Patient(String name, String address, int age,long mobile) {
         super(name, address, age,mobile);
     }
 }
