+++ src/com/company/Entity/Visitor.java
@@ -1,7 +1,7 @@
 package com.company.Entity;
 
 public class Visitor extends User{
-    public Visitor(String name, String address, int age,int mobile) {
+    public Visitor(String name, String address, int age,long mobile) {
         super(name, address, age,mobile);
     }
 }
